refactor(media): api methods in sperate file
